Azure_working_function/requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
azure-identity
azure-storage-blob
azure-storage-queue
paramiko
azure-keyvault-secrets
xmltodict
azure-storage-common
